// apps/qa-lead/src/app/qa-lead/data-access/model/ui/qa-lead-item-general-info.ts
import { ItemTransitionStatus } from "@shared/data-access/model/api/enum/item-transition-status";

export type QaLeadItemGeneralInfo = QaItemGeneralInfo & {
  isQaSuccessAll: boolean;
  isSockPrint: boolean;
};

export type QaItemGeneralInfo = {
  isPriority: boolean;
  partnerId: string;
  orderId: number;
  partnerOrderId: string;
  barcode: string;
  status: ItemTransitionStatus | null;
  statusName: string;
  sku: string;
  cooName: string | null;
  isAllowChangeCoo: boolean;
  quantity: number;
  slaDateOnUtc: Date;
  isJit: boolean;
  preQrCode: string | null;
};